requirements/local.txt
-r base.txt

coverage==3.7.1
django-debug-toolbar==1.4.0
mock==1.3.0
factory_boy